+++ src/components/content/Content.form.js
@@ -1,12 +1,17 @@
 import baseEditForm from '../base/Base.form';
 
 import ContentEditDisplay from './editForm/Content.edit.display';
+import ContentEditLogic from './editForm/Content.edit.logic';
 
 export default function(...extend) {
   return baseEditForm([
     {
       key: 'display',
-      components: ContentEditDisplay
-    }
+      components: ContentEditDisplay,
+    },
+    {
+      key: 'logic',
+      components: ContentEditLogic,
+    },
   ], ...extend);
 }